requirements.txt
fastapi[standard]
sqlmodel
bcrypt
pyjwt
alembic
pymysql